NEW: Prune ends of exact matches
This makes pruning for inexact matches much stronger:
when an exact match is pruned, the neighbouring inexact matches are now
also removed.

Drawback: Where the number of matches is large, pruning can now become
quite slow, because the pruned area loses all its structure. We could
re-introduce the prune factor, but that is for later.